[microvm] E: Merge Code for MicroVM and Hyperlight
The I/O thread code can be exactly the same. For now, hyperlight doesn't
use every channel that's created, so they're left with a leading
underscore.

The memory thread is not exactly the same. A closure is used to
encapsulate the different code.
