+++ src/microvm/src/io.rs
@@ -42,7 +42,46 @@ pub struct IoThread {
     incoming: VecDeque<Message>,
     /// Queue of outgoing messages.
     outgoing: VecDeque<Message>,
+    /// Command sender to the VMM.
+    _vmm_control_tx: Sender<ControlCommand>,
+    /// Response receiver from the VMM.
+    vmm_control_rx: Receiver<ControlCommandResponse>,
+    // TODO: channels to an outside issuer of snapshot commands and to linuxd.
 }
+
+//==================================================================================================
+// Enums
+//==================================================================================================
+
+///
+/// # Description
+///
+/// Control plane commands.
+///
+#[derive(PartialEq)]
+pub enum ControlCommand {
+    _StartMicroVm,
+    _LoadSnapshotAndRun,
+    _PauseMicroVm,
+    _PauseAndCreateSnapshot,
+    _CreateSnapshot,
+    _ResumeMicroVm,
+    LinuxDaemonFlushed,
+}
+
+///
+/// # Description
+///
+/// Control plane command responses.
+///
+#[derive(PartialEq)]
+pub enum ControlCommandResponse {
+    MicroVmPaused,
+    SnapshotCreated,
+    FlushOutput,
+    FlushInput,
+}
+
 //==================================================================================================
 // Implementations
 //==================================================================================================
@@ -58,6 +97,8 @@ impl IoThread {
     /// - `gateway`: Connection to gateway.
     /// - `microvm_rx`: MicroVM receiver.
     /// - `microvm_tx`: MicroVM sender.
+    /// - `vmm_control_tx`: Command sender.
+    /// - `vmm_control_rx`: Response receiver.
     ///
     /// # Returns
     ///
@@ -67,9 +108,12 @@ impl IoThread {
         gateway: Gateway,
         microvm_rx: Receiver<Message>,
         microvm_tx: Sender<Message>,
+        vmm_control_tx: Sender<ControlCommand>,
+        vmm_control_rx: Receiver<ControlCommandResponse>,
     ) -> JoinHandle<Result<()>> {
         thread::spawn(move || {
-            let mut io_thread: IoThread = IoThread::new(gateway, microvm_rx, microvm_tx)?;
+            let mut io_thread: IoThread =
+                IoThread::new(gateway, microvm_rx, microvm_tx, vmm_control_tx, vmm_control_rx)?;
             io_thread.run()?;
             Ok(())
         })
@@ -85,6 +129,8 @@ impl IoThread {
     /// - `gateway`: Connection to gateway.
     /// - `microvm_rx`: MicroVM receiver.
     /// - `microvm_tx`: MicroVM sender.
+    /// - `vmm_control_tx`: Command sender.
+    /// - `vmm_control_rx`: Response receiver.
     ///
     /// # Returns
     ///
@@ -94,6 +140,8 @@ impl IoThread {
         gateway: Gateway,
         microvm_rx: Receiver<Message>,
         microvm_tx: Sender<Message>,
+        vmm_control_tx: Sender<ControlCommand>,
+        vmm_control_rx: Receiver<ControlCommandResponse>,
     ) -> Result<Self> {
         Ok(Self {
             gateway,
@@ -101,33 +149,27 @@ impl IoThread {
             microvm_tx,
             incoming: VecDeque::new(),
             outgoing: VecDeque::new(),
+            _vmm_control_tx: vmm_control_tx,
+            vmm_control_rx,
         })
     }
 
     ///
     /// # Description
     ///
-    /// Runs the I/O thread.
+    /// Runs the I/O thread according to the state in the snapshotting protocol state machine.
     ///
     /// # Returns
     ///
     /// Upon success, empty is returned. Otherwise, an error is returned instead.
     ///
     fn run(&mut self) -> Result<()> {
-        let mut round: usize = 0;
-
-        // Cycle through actions to avoid starvation.
         loop {
-            if round % 4 == 0 {
-                self.try_receive_from_microvm()?;
-            } else if round % 4 == 1 {
-                self.try_send_to_gateway()?;
-            } else if round % 4 == 2 {
-                self.try_receive_from_gateway()?;
-            } else {
-                self.try_send_to_microvm()?;
-            }
-            round += 1;
+            self.try_receive_from_microvm()?;
+            self.try_send_to_gateway()?;
+            self.try_receive_from_gateway()?;
+            self.try_send_to_microvm()?;
+            self.try_receive_from_vmm_control()?;
         }
     }
 
@@ -138,17 +180,18 @@ impl IoThread {
     ///
     /// # Returns
     ///
-    /// Upon success, the received message is returned. Otherwise, an error is returned.
+    /// Upon success, the received message is pushed into the `incoming` queue, and `true` is returned.
+    /// Otherwise, if it would block, `false` is returned. Otherwise, an error is returned.
     ///
-    fn try_receive_from_gateway(&mut self) -> Result<()> {
+    fn try_receive_from_gateway(&mut self) -> Result<bool> {
         match self.gateway.try_receive() {
             Ok(message) => {
                 self.incoming.push_back(message);
-                Ok(())
+                Ok(true)
             },
             Err(e) => {
                 if e.kind() == ErrorKind::WouldBlock {
-                    Ok(())
+                    Ok(false)
                 } else {
                     let reason: String =
                         format!("failed to receive message from the gateway (error={e:?})");
@@ -166,9 +209,10 @@ impl IoThread {
     ///
     /// # Returns
     ///
-    /// Upon success, the received message is returned. Otherwise, an error is returned.
+    /// Upon success, the received message is pushed into the `outgoing` queue, and `true`is returned.
+    /// Otherwise, if the channel is empty, `false` is returned. Otherwise, an error is returned.
     ///
-    fn try_receive_from_microvm(&mut self) -> Result<()> {
+    fn try_receive_from_microvm(&mut self) -> Result<bool> {
         match self.microvm_rx.try_recv() {
             Ok(mut message) => {
                 profiler::timestamp_message!(
@@ -177,9 +221,9 @@ impl IoThread {
                         + std::mem::offset_of!(syscall::unistd::message::WriteRequest, buffer)
                 );
                 self.outgoing.push_back(message);
-                Ok(())
+                Ok(true)
             },
-            Err(TryRecvError::Empty) => Ok(()),
+            Err(TryRecvError::Empty) => Ok(false),
             Err(TryRecvError::Disconnected) => {
                 let reason: String = "the microvm has disconnected".to_string();
                 // When the guest finishes , the vCPU thread will disconnect from this thread. This
@@ -250,5 +294,64 @@ impl IoThread {
             },
             None => Ok(()),
         }
+    }
+
+    ///
+    /// # Description
+    ///
+    /// Attempts to receive a response from the VMM.
+    ///
+    /// # Returns
+    ///
+    /// Upon success, empty is returned. Otherwise, an error is returned.
+    ///
+    fn try_receive_from_vmm_control(&mut self) -> Result<()> {
+        match self.vmm_control_rx.try_recv() {
+            Ok(response) => match response {
+                ControlCommandResponse::FlushOutput => self.flush_microvm_output(),
+                ControlCommandResponse::FlushInput => self.flush_linuxd_input(),
+                _ => Ok(()), // TODO: forward to whoever is interested. This requires having control channels.
+            },
+            Err(TryRecvError::Empty) => Ok(()),
+            Err(TryRecvError::Disconnected) => {
+                let reason: String = "the vmm has disconnected".to_string();
+                // When the guest finishes , the vCPU thread will disconnect from this thread. This
+                // situation is normal and should not create an error log.
+                anyhow::bail!(reason)
+            },
+        }
+    }
+
+    /// # Description
+    ///
+    /// Attempts to flush all outstanding output from the MicroVM to the Linux daemon.
+    ///
+    /// # Returns
+    ///
+    /// Upon success, empty is returned. Otherwise, an error is returned.
+    ///
+    fn flush_microvm_output(&mut self) -> Result<()> {
+        while self.try_receive_from_microvm()? {
+            // Keep looping until `microvm_rx` is empty, which breaks the loop.
+        }
+        while !self.outgoing.is_empty() {
+            self.try_send_to_gateway()?;
+        }
+        Ok(())
+    }
+
+    /// # Description
+    ///
+    /// Attempts to flush all outstanding input from the the Linux daemon to the MicroVM.
+    ///
+    /// # Returns
+    ///
+    /// Upon success, empty is returned. Otherwise, an error is returned.
+    ///
+    fn flush_linuxd_input(&mut self) -> Result<()> {
+        while self.try_receive_from_gateway()? {
+            // Keep looping until receiving from the gateway would block, which breaks the loop.
+        }
+        Ok(())
     }
 }
